Extracted OrganizationItem
