+++ frontend/src/components/organizationProfile/organizationProfile.js
@@ -2,13 +2,18 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { withStyles, createStyleSheet } from 'material-ui/styles';
 import Grid from 'material-ui/Grid';
+import Paper from 'material-ui/Paper';
 import Typography from 'material-ui/Typography';
 import MainContentWrapper from '../../components/common/mainContentWrapper';
-import HqProfile from './hqProfile';
+import OrganizationItem from './organizationItem';
 import CountryOfficesList from './countryOfficesList';
 
 const messages = {
   header: 'Organization Profile',
+};
+
+const hgProfileMockData = {
+  users: 25, update: '01 Jan 2016',
 };
 
 const styleSheet = createStyleSheet('OrganizationProfile', (theme) => {
@@ -42,7 +47,9 @@ const OrganizationProfile = (props) => {
       <MainContentWrapper>
         <Grid container direction="column" gutter={40}>
           <Grid item>
-            <HqProfile />
+            <Paper>
+              <OrganizationItem users={hgProfileMockData.users} update={hgProfileMockData.update} />
+            </Paper>
           </Grid>
           <Grid item>
             <CountryOfficesList />
